created - update database class #471
