+++ accessibility-checker.php
@@ -132,7 +132,6 @@ require_once plugin_dir_path( __FILE__ ) . 'includes/purge.php';
 /**
  * Filters and Actions
  */
-add_action( 'admin_init', 'edac_update_database', 10 );
 add_action( 'add_meta_boxes', 'edac_register_meta_boxes' );
 add_action( 'admin_menu', 'edac_add_options_page' );
 add_action( 'admin_init', 'edac_register_setting' );
@@ -144,52 +143,6 @@ add_action( 'template_redirect', 'edac_before_page_render' );
 if ( is_plugin_active( 'oxygen/functions.php' ) ) {
 	add_action( 'added_post_meta', 'edac_oxygen_builder_save_post', 10, 4 );
 	add_action( 'updated_post_meta', 'edac_oxygen_builder_save_post', 10, 4 );
-}
-
-/**
- * Create/Update database
- *
- * @return void
- */
-function edac_update_database() {
-
-	global $wpdb;
-	$table_name = $wpdb->prefix . 'accessibility_checker';
-
-	$query = $wpdb->prepare( 'SHOW TABLES LIKE %s', $wpdb->esc_like( $table_name ) );
-	// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching -- Prepare above, Safe variable used for table name, caching not required for one time operation.
-	if ( get_option( 'edac_db_version' ) !== EDAC_DB_VERSION || $wpdb->get_var( $query ) !== $table_name ) {
-
-		$charset_collate = $wpdb->get_charset_collate();
-		$sql             = "CREATE TABLE $table_name (
-			id bigint(20) NOT NULL AUTO_INCREMENT,
-			postid bigint(20) NOT NULL,
-			siteid text NOT NULL,
-			type text NOT NULL,
-			rule text NOT NULL,
-			ruletype text NOT NULL,
-			object mediumtext NOT NULL,
-			recordcheck mediumint(9) NOT NULL,
-			created timestamp NOT NULL default CURRENT_TIMESTAMP,
-			user bigint(20) NOT NULL,
-			ignre mediumint(9) NOT NULL,
-			ignre_global mediumint(9) NOT NULL,
-			ignre_user bigint(20) NULL,
-			ignre_date timestamp NULL,
-			ignre_comment mediumtext NULL,
-			UNIQUE KEY id (id),
-			KEY postid_index (postid)
-		) $charset_collate;";
-
-		require_once ABSPATH . 'wp-admin/includes/upgrade.php';
-		dbDelta( $sql );
-
-	}
-
-	// Update database version option.
-	$option_name = 'edac_db_version';
-	$new_value   = EDAC_DB_VERSION;
-	update_option( $option_name, $new_value );
 }
 
 /**
